separate zui theme in zui.tsx to allow testing of zui components
Restore old theme usage in index.tsx
